split code to header files
