+++ src/main.cpp
@@ -1,210 +1,16 @@
 #include <Adafruit_MotorShield.h>  // https://github.com/adafruit/Adafruit_Motor_Shield_V2_Library
 #include <Arduino.h>
-#include <LinkedList.h>                          // https://github.com/ivanseidel/LinkedList
 #include <Time.h>                                // standard Arduino time library
 #include <Wire.h>                                // standard Arduino i2c library
 #include <avr/wdt.h>                             // Arduino watchdog library
 #include <stdint.h>                              // integer definitions: int8_t, int16_t, ..., uint8_t, ...
 #include "DS3232RTC.h"                           // https://github.com/JChristensen/DS3232RTC
-#include "utility/Adafruit_MS_PWMServoDriver.h"  // https://github.com/adafruit/Adafruit_Motor_Shield_V2_Library
+// #include "utility/Adafruit_MS_PWMServoDriver.h"  // https://github.com/adafruit/Adafruit_Motor_Shield_V2_Library
+#include "DosingPump.h"
 
 #define SERIAL_SPEED = 9600;
 
 typedef Adafruit_MotorShield MotorShield;
-typedef Adafruit_DCMotor DcMotor;
-
-typedef struct {
-    uint8_t startHour;
-    uint8_t startMinute;
-    uint8_t doseMiliLiters;
-} DosingTask;
-
-class DosingSchedule {
-    LinkedList<DosingTask*> dosingSchedule;  // keep a list of dosing task pointers
-
-    bool isValidHourMinute(uint8_t hour, uint8_t minute) {
-        if ((hour < 24) && (minute < 60)) {
-            return true;
-        }
-        return false;
-    }
-
-   public:
-    DosingSchedule() {
-        dosingSchedule = LinkedList<DosingTask*>();
-    }
-
-    uint8_t size() {
-        return dosingSchedule.size();
-    }
-
-    DosingTask* getTaskAtIndex(uint8_t index) {
-        return dosingSchedule.get(index);
-    }
-
-    DosingTask* getTaskAtHourMinute(uint8_t hour, uint8_t minute) {
-        // returns the first match, will never return any duplicates if they exist
-
-        DosingTask* dosingTaskPointer;
-
-        for (int i = 0; i < dosingSchedule.size(); i++) {
-            if ((dosingTaskPointer->startHour == hour) && (dosingTaskPointer->startMinute == minute)) {
-                return dosingTaskPointer;
-            }
-        }
-        return nullptr;
-    }
-
-    bool addTask(uint8_t startHour, uint8_t startMinute, uint8_t doseMiliLiters) {
-        // add new task at the end of the list
-
-        if (doseMiliLiters < 1) return false;
-
-        // check startHour and startMinute are valid
-        // check there is no other dosing task at the same time
-        if (isValidHourMinute(startHour, startMinute) && (getTaskAtHourMinute(startHour, startMinute) == nullptr)) {
-            DosingTask* dosingTaskPointer = new DosingTask();
-            dosingTaskPointer->startHour = startHour;
-            dosingTaskPointer->startMinute = startMinute;
-            dosingTaskPointer->doseMiliLiters = doseMiliLiters;
-            dosingSchedule.add(dosingTaskPointer);
-            return true;
-        }
-        return false;
-    }
-
-    bool updateTask(uint8_t index, uint8_t startHour, uint8_t startMinute, uint8_t doseMiliLiters) {
-        // update existing task at index
-
-        if (doseMiliLiters < 1) return false;
-
-        DosingTask* dosingTaskPointer = getTaskAtIndex(index);
-
-        // check the index is valid, not NULL
-        // check startHour and startMinute are valid
-        // check there is no other dosing task at the same time
-        if ((dosingTaskPointer != nullptr) && isValidHourMinute(startHour, startMinute) && (getTaskAtHourMinute(startHour, startMinute) == nullptr)) {
-            dosingTaskPointer->startHour = startHour;
-            dosingTaskPointer->startMinute = startMinute;
-            dosingTaskPointer->doseMiliLiters = doseMiliLiters;
-            return true;
-        }
-        return false;
-    }
-
-    void removeTask(uint8_t index) {
-        // remove task from schedule 
-
-        DosingTask* dosingTaskPointer = dosingSchedule.remove(index);
-        delete dosingTaskPointer;
-    }
-};
-
-/*!
-    @brief To calibrate the dosing pump, call `startCalibration()`, 
-    wait for the pump to dispense 100 mL, and call `stopCalibration()`.
-     - min dose is 1 mL
-     - max dose is 255 mL
-     - dose increments by 1mL. 
-*/
-class DosingPump {
-    MotorShield motorShiled;
-    uint8_t motorPort;
-    DcMotor* dosingPumpPointer;
-    DosingSchedule dosingSchedule;
-
-    unsigned long startPumpingMilis;  // used in the `loop()` method
-    unsigned long startCalibrationMilis;
-    unsigned long dosingPeriodMilis;
-    // default value, should be overwirten by calibration
-    // must persist calibrated value to eeprom
-    unsigned long milisPerMiliLiter = 1000;
-
-    DosingTask* dosingTaskPointer;  // used in the `loop()` method
-
-    enum State {
-        IDLE = 0,
-        PUMPING = 1,
-        CALIBRATING = 2
-    } state;
-
-    unsigned long miliLitersToMilis(uint8_t miliLiters) {
-        return milisPerMiliLiter * miliLiters;
-    }
-
-   public:
-    DosingPump(Adafruit_MotorShield attachToShield, uint8_t attachToMotorPort) : motorShiled(attachToShield),
-                                                                                 motorPort(attachToMotorPort) {
-    }
-
-    ~DosingPump() {
-        delete dosingPumpPointer;
-        delete dosingTaskPointer;
-    }
-
-    void setSpeed(uint8_t speed) {
-        dosingPumpPointer->setSpeed(speed);
-    }
-
-    void startCalibration() {
-        startCalibrationMilis = millis();
-        startPumping();
-        state = CALIBRATING;
-    }
-
-    void stopCalibration() {
-        /* should be called after the pump has dispensed 100mL */
-        milisPerMiliLiter = (unsigned long)(millis() - startCalibrationMilis) / 100;
-        stopPumping();
-        state = IDLE;
-    }
-
-    void startPumping() {
-        /* change wires polarity to reverse direction, or use BACKWARD */
-        dosingPumpPointer->run(FORWARD);
-    }
-
-    void stopPumping() {
-        dosingPumpPointer->run(RELEASE);
-    }
-
-    void setup() {
-        dosingPumpPointer = motorShiled.getMotor(motorPort);
-        setSpeed(240);
-        state = IDLE;
-
-        // todo: load configuration (milisPerMiliLiter, DosingSchedule)
-        dosingSchedule = DosingSchedule();
-        dosingSchedule.addTask(12, 30, 3);  // hour, minute, doseMiliLiters
-        dosingSchedule.addTask(14, 30, 3);
-        dosingSchedule.addTask(16, 30, 3);
-    }
-
-    void loop() {
-        switch (state) {
-            case IDLE:
-                dosingTaskPointer = dosingSchedule.getTaskAtHourMinute(hour(), minute());
-                if (dosingTaskPointer != nullptr) {
-                    dosingPeriodMilis = miliLitersToMilis(dosingTaskPointer->doseMiliLiters);
-                    startPumpingMilis = millis();
-                    startPumping();
-                    state = PUMPING;
-                }
-                break;
-
-            case PUMPING:
-                if ((unsigned long)(millis() - startPumpingMilis) > dosingPeriodMilis) {
-                    stopPumping();
-                    state = IDLE;
-                }
-                break;
-
-            case CALIBRATING:
-                // do not mess with the calibrating methods
-                break;
-        }
-    }
-};
 
 class DosingStation {
     unsigned long startSleepMilis;
